requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.9
marshmallow==3.19.0
moviepy==1.0.3
pydub==0.25.1
Werkzeug==2.2.3
python-dotenv==1.0.0
celery==5.2.7
redis==4.5.4
gunicorn==22.0.0
python-magic==0.4.27
rq==1.11.1
redis==4.5.4
Flask-Limiter==3.3.0